requirements.txt
pandas
numpy
matplotlib
duckdb
openai
google-generativeai
